Moved stuff and renamed resources to match new ID
