+++ src/main/java/me/hydos/J64/minecraft/block/TVBlock.java
@@ -1,4 +1,4 @@
-package me.hydos.J64.minecraft.blocks;
+package me.hydos.J64.minecraft.block;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
