+++ src/main/resources/assets/jcraft64/models/block/n64_charcoal.bbmodel
