--- src/main/resources/assets/jcraft64/models/item/n64_black.json
@@ -1,3 +1,0 @@
-{
-  "parent": "jcraft64:block/n64_black"
-}
