+++ src/main/resources/assets/jcraft64/models/item/n64_charcoal.json
@@ -1,0 +1,3 @@
+{
+  "parent": "jcraft64:block/n64_charcoal"
+}
