+++ src/main/resources/assets/jcraft64/models/item/tv.json
@@ -1,0 +1,3 @@
+{
+  "parent": "jcraft64:block/tv"
+}
