+++ src/main/resources/assets/jcraft64/textures/block/charcoal_n64_bottom.png
